硬解码seek优化完成

+++ app/src/main/cpp/Tool/XData.h
@@ -18,7 +18,7 @@ struct XData {
     unsigned char *framedatas[8]={0};
     int size=0;
     bool isAudio=false;
-    int pts=0;
+    long pts=0;
     int dts=0;
     bool Alloc(int size,const char *d=0);
     void Drop();
